+++ packages/@livestore/adapter-web/src/web-worker/client-session/persisted-adapter.ts
@@ -203,6 +203,7 @@ export const makePersistedAdapter =
       yield* Queue.offer(bootStatusQueue, { stage: 'loading' })
 
       const sqlite3 = yield* Effect.promise(() => loadSqlite3())
+      const makeSqliteDb = sqliteDbFactory({ sqlite3 })
 
       const LIVESTORE_TAB_LOCK = `livestore-tab-lock-${storeId}`
       const LIVESTORE_SHARED_WORKER_TERMINATION_LOCK = `livestore-shared-worker-termination-lock-${storeId}`
@@ -233,11 +234,10 @@ export const makePersistedAdapter =
       // we're here trying to get the snapshot directly from storage
       // we usually speeds up the boot process by a lot.
       // We need to be extra careful though to not run into any race conditions or inconsistencies.
-      // TODO also verify persisted data
       const dataFromFile =
         options.experimental?.disableFastPath === true || opfsWarning !== undefined
           ? undefined
-          : yield* readPersistedStateDbFromClientSession({ storageOptions, storeId, schema }).pipe(
+          : yield* readPersistedStateDbFromClientSession({ storageOptions, storeId, schema, makeSqliteDb }).pipe(
               Effect.tapError((error) =>
                 Effect.logDebug('[@livestore/adapter-web:client-session] Could not read persisted state db', error, {
                   storeId,
@@ -432,31 +432,27 @@ export const makePersistedAdapter =
 
       // TODO maybe bring back transfering the initially created in-memory db snapshot instead of
       // re-exporting the db
-      const initialResult =
+      const { sqliteDb, snapshotByteLength, migrationsReport } =
         dataFromFile === undefined
-          ? yield* runInWorker('GetRecreateSnapshot', sharedWorker.GetRecreateSnapshot({})).pipe(
-              Effect.map(({ snapshot, migrationsReport }) => ({
-                _tag: 'from-leader-worker' as const,
-                snapshot,
-                migrationsReport,
-              })),
-            )
-          : { _tag: 'fast-path' as const, snapshot: dataFromFile }
-
-      const migrationsReport =
-        initialResult._tag === 'from-leader-worker' ? initialResult.migrationsReport : { migrations: [] }
-
-      const makeSqliteDb = sqliteDbFactory({ sqlite3 })
-      const sqliteDb = yield* makeSqliteDb({ _tag: 'in-memory' })
-
-      sqliteDb.import(initialResult.snapshot)
+          ? yield* Effect.gen(function* () {
+              const { snapshot, migrationsReport } = yield* runInWorker(
+                'GetRecreateSnapshot',
+                sharedWorker.GetRecreateSnapshot({}),
+              )
+              const sqliteDb = yield* Effect.acquireRelease(makeSqliteDb({ _tag: 'in-memory' }), (db) =>
+                Effect.sync(() => db.close()),
+              )
+              sqliteDb.import(snapshot)
+              return { sqliteDb, snapshotByteLength: snapshot.byteLength, migrationsReport }
+            })
+          : { ...dataFromFile, migrationsReport: { migrations: [] } }
 
       const numberOfTables =
         sqliteDb.select<{ count: number }>(`select count(*) as count from sqlite_master`)[0]?.count ?? 0
       if (numberOfTables === 0) {
         return yield* UnknownError.make({
           cause: `Encountered empty or corrupted database`,
-          payload: { snapshotByteLength: initialResult.snapshot.byteLength, storageOptions: options.storage },
+          payload: { snapshotByteLength, storageOptions: options.storage },
         })
       }